+++ CDK/src/test/java/org/familydirectory/cdk/test/cognito/FamilyDirectoryCognitoStackTest.java
@@ -122,10 +122,9 @@ class FamilyDirectoryCognitoStackTest {
                                                                                                                                                                       "CONFIRM_WITH_LINK",
                                                                                                                                                                       "EmailMessageByLink",
                                                                                                                                                                       "Verify your account by" +
-                                                                                                                                                                      " clicking on {##Verify" +
-                                                                                                                                                                      " Email##}", "EmailSubjectByLink",
-                                                                                                                                                                      "Verify your new " +
-                                                                                                                                                                      "account")))));
+                                                                                                                                                                      " clicking on {##Verify " +
+                                                                                                                                                                      "Email##}", "EmailSubjectByLink",
+                                                                                                                                                                      "Verify your new account")))));
         assertTrue(sesArnCapture.asObject()
                                 .toString()
                                 .contains("ses:%s:%s:identity/%s".formatted(FamilyDirectoryCdkApp.DEFAULT_REGION, FamilyDirectoryCdkApp.DEFAULT_ACCOUNT, FamilyDirectoryDomainStack.HOSTED_ZONE_NAME)));
@@ -140,9 +139,9 @@ class FamilyDirectoryCognitoStackTest {
                                                                                                 entry("AllowedOAuthScopes", List.of("email", "openid", "aws.cognito.signin.user.admin", "profile")),
                                                                                                 entry("CallbackURLs",
                                                                                                       singletonList(FamilyDirectoryCdkApp.HTTPS_PREFIX + FamilyDirectoryDomainStack.HOSTED_ZONE_NAME)),
-                                                                                                entry("ExplicitAuthFlows", List.of("ALLOW_USER_SRP_AUTH", "ALLOW_REFRESH_TOKEN_AUTH")),
                                                                                                 entry("LogoutURLs",
                                                                                                       singletonList(FamilyDirectoryCdkApp.HTTPS_PREFIX + FamilyDirectoryDomainStack.HOSTED_ZONE_NAME)),
+                                                                                                entry("ExplicitAuthFlows", List.of("ALLOW_USER_SRP_AUTH", "ALLOW_REFRESH_TOKEN_AUTH")),
                                                                                                 entry("GenerateSecret", FamilyDirectoryCognitoStack.COGNITO_USER_POOL_CLIENT_GENERATE_SECRET),
                                                                                                 entry("PreventUserExistenceErrors", "ENABLED"),
                                                                                                 entry("ReadAttributes", List.of("email", "email_verified")), entry("SupportedIdentityProviders",